Improve struct constructor
Check that the constructor arguments specified container is the direct
container of specified path.  If not, attempt to determine the proper
container using the tree from the specified container.

In the future I may eliminate "container" as used by struct and replace it
with "context" property which may be even more flexible.